Fix LDP relocation to not include a constant offset
The LDP instruction takes the base as PC & ~0x7. This was implemented
as a -4 when the PC was not 8 byte aligned. This is a problem both
because when relaxing code the PC could change and hence the offset
become invalid and also the address of the destination was impossible
to calculate as a symbol+offset address was impossible to distinguish
from symbol+offset-4.

The adjustment for the PC happens at patch time now.