Regression fix: force ABI/architecture for tests.
The default ABI is set during configuration based on the target name
and wrong ABI is picked by the assembler when e.g. assembling for -march=mips32*
without specifying -mabi for mips64-*-* target.

In MSA case, the dump for matching was generated for mips1 arch and with
mips64 tool it will fail.