Add MIPS32 FPU64 GDB target descriptions
Add a couple of new target descriptions for MIPS32 targets with 64-bit
floating point registers, with and without DSP.

These are identical to their 32-bit FPU counterparts except that the FP
registers are 64-bits long to allow for debugging of F64=1 MIPS32 targets
such as MIPS32r2 compatible cores, and they include the Config5 CP0
register which has an FRE bit.

F64 targets have an FR bit in the status register to specify the effective
register size, and the FRE bit is used with FR=1 to enable a compatibility
mode which is similar to FR=0 but with usable odd doubles which don't
alias with any other FP registers.

FIXME update changelogs

gdb/ChangeLog:

	* features/Makefile (WHICH): Add mips-fpu64-linux and
	mips-fpu64-dsp-linux.
	(mips-fpu64-expedite): Set.
	(mips-fpu64-dsp-expedite): Set.
	* features/mips-cp0-fpu64.xml: New file.
	* features/mips-fpu64-dsp-linux.xml: New file.
	* features/mips-fpu64-linux.xml: New file.
	* features/mips-fpu64.xml: New file.
	* features/mips-fpu64-dsp-linux.c: New generated file.
	* features/mips-fpu64-linux.c: New generated file.
	* regformats/mips-fpu64-dsp-linux.dat: New generated file.
	* regformats/mips-fpu64-linux.dat: New generated file.

gdb/gdbserver/ChangeLog:

	* Makefile.in (clean): Delete mips-fpu64-linux.c and
	mips-fpu64-dsp-linux.c.
	(mips-fpu64-linux.c): New rule.
	(mips-fpu64-dsp-linux.c): New rule.
	* configure.srv (srv_regobj): Add mips-fpu64-linux.o and
	mips-fpu64-dsp-linux.o.
	(srv_xmlfiles): Add mips-fpu64-linux.xml,
	mips-fpu64-dsp-linux.xml, mips-cp0-fpu64.xml and mips-fpu64.xml.