MIPS: Add config5 to MIPS GDB target descriptions
Add the config5 register to the MIPS and MIPS64 GDB target description
files and regenerate various dependent files. This register contains the
FRE bit which is used with the FR bit in the status register to implement
a new floating point register mode.

The register is placed with the other cp0 registers at register number 38
(which used to be the first FP register number). The explicit FPU and DSP
register numbers are removed so that they can all shuffle up by one to
give space for config5. The explicit FPU register number used to be
necessary due to the PC register being defined non-linearly at register
number 37, but the explicit DSP register numbers didn't seem to serve a
particular purpose as far as I can tell.

An corresponding -1 entry is added to mips_base_regs in
gdbserver/linux-mips-low.c for config5, indicating that it cannot be
accessed using PTRACE_PEEK/PTRACE_POKE.

gdb/ChangeLog:

	* features/mips-cp0.xml: Add config5 register.
	* features/mips64-cp0.xml: Likewise.
	* features/mips-fpu.xml: Remove explicit register number.
	* features/mips64-fpu.xml: Likewise.
	* features/mips-dsp.xml: Remove explicit register numbers.
	* features/mips64-dsp.xml: Likewise.
	* features/mips-dsp-linux.c: Regenerate.
	* features/mips64-dsp-linux.c: Likewise.
	* features/mips-linux.c: Likewise.
	* features/mips64-linux.c: Likewise.
	* regformats/mips-dsp-linux.dat: Regenerate.
	* regformats/mips64-dsp-linux.dat: Likewise.
	* regformats/mips-linux.dat: Likewise.
	* regformats/mips64-linux.dat: Likewise.

gdb/gdbserver/ChangeLog:

	* linux-mips-low.c (mips_num_regs, mips_dsp_num_regs): Increment.
	(mips_base_regs): Add -1 entry for Config5.