MIPS: Enhance cooked FP format
Define cooked FP types as unions representing the possible interpretations
of the registers.

Previously the type of the cooked floating point registers depended
directly on the raw type, so for MIPS32 targets they were single
precision, and for MIPS64 targets they were single or double precision
depending on the FP mode.

However in both cases it is possible for the FP registers to be
interpreted as either single or double precision values depending on the
instruction executed, and in a different way depending on the FP mode.

When FR=0:
- Even FP registers can be interpreted as a double (aliasing both odd and
  even singles), or a single (aliasing the least significant half of the
  double).
- Odd FP registers can be interpreted only as a single (aliasing the most
  significant half of the even double).

When FR=1:
- All FP registers can be interpreted as a double, or a single (aliasing
  the least significant half of the double).

Some instructions can also treat them as 32-bit or 64-bit integers for the
purposes of converting between formats.

Therefore define two cooked fp types, as follows.

The first type, fp32, is used for odd FP registers when FR=0, which can
only be interpreted as singles or 32-bit integers. It is equivalent to
this C definition:
  union fp32 {
    float f32;
    int32_t i32;
  };

The second type, fp64, is used for even FP registers when FR=0, and all FP
registers when FR=1, which can be interpreted as either singles or
doubles, with the single always in the least significant half of the
double. It is roughly equivalent to this C definition (when interpreted as
little endian):
  union fp64 {
    float  f32;
    double f64;
    int32  i32;
    int64  i64;
  };

Since multiple bits of code need to handle the different raw register
layouts and these new cooked types for different modes, the details are
abstracted with the following functions:
- mips_get_fp_single_location: Gets the raw register and offset containing
  a single precision register value.
- mips_get_fp_double_location: Gets the raw register(s) and offset(s)
  composing a double precision register value.
- mips_get_fp_multi_location: Gets the raw register(s) and offset(s)
  composing a cooked register value.

These abstractions are used by the code which reads and writes the cooked
(pseudo) FP registers, and the code which prints out the FP registers
("info float").

Now that the cooked type may be composed of multiple raw registers (where
only single precision raw values are provided), there is no need to
provide conversion functions from the 32-bit cooked value to double.
Instead the opposite is required for extracting a single out of the 64-bit
cooked value.

For example, with FR=0:
  (gdb) info float
  ...
  f30: 0x41f00000 flt: 30                dbl: 6442451998.9999998
  f31: 0x41f80000 flt: 31
  (gdb) p $f30
  $1 = {f32 = 30, f64 = 6442451998.9999998,
        i32 = 1106247680, i64 = 4753549407795806208}
  (gdb) p $f31
  $2 = {f32 = 31, i32 = 1106771968}

And with FR=1:
  (gdb) info float
  ...
  f30: 0x403e000041f00000 flt: 30                dbl: 30.000003930181265
  f31: 0x403f000041f80000 flt: 31                dbl: 31.00000393204391
  (gdb) p $f30
  $1 = {f32 = 30, f64 = 30.000003930181265,
        i32 = 1106247680, i64 = 4629137468089696256}
  (gdb) p $f31
  $2 = {f32 = 31, f64 = 31.00000393204391,
        i32 = 1106771968, i64 = 4629418943066931200}

gdb/ChangeLog:

	* mips-tdep.c (mips_register_reggroup_p): Match FP registers with
	mips_float_register_p rather than TYPE_CODE_FLT.
	(struct mips_reg_part, mips_get_fp_single_location,
	mips_get_fp_double_location, mips_get_fp_multi_location,
	mips_regcache_raw_read_parts, mips_regcache_raw_write_parts,
	mips_fp32_type, mips_fp64_type, mips_fp_type): New definitions.
	(mips_pseudo_register_read, mips_pseudo_register_write,
	mips_convert_register_float_case_p, mips_register_to_value,
	mips_value_to_register, mips_register_type,
	mips_pseudo_register_type, mips_read_fp_register_single,
	mips_read_fp_register_double, mips_print_fp_register): Convert to
	handle new cooked FP format.
	(mips_gdbarch_init): Initialise tdep->fp32_type and
	tdep->fp64_type.
	mips-tdep.h (struct gdbarch_tdep): Add fp32_type and fp64_type.