MIPS: Add support for hybrid fp32/fp64 mode
Add support for the new hybrid fp32/fp64 mode, which is intended to allow
fp32 (which requires FR=0) and MSA code (which requires FR=1) to coexist
in the same process. It is determined by the FRE bit in the Config5
register, which causes single-precision instructions to trap so that the
kernel can emulate them using the FR=0 aliasing of singles in even
doubles.

When FR=1 and FRE=1:
- Even FP registers can be interpreted as a double, or a single (aliasing
  the least significant half of the double).
- Odd FP registers can be interpreted as a double, or a single (aliasing
  the most significant half of the even double).

The even registers can use the existing fp64 type, however a new type is
required for the odd FP registers when FRE=1, since the single and double
interpretations don't alias. It is equivalent to this C definition:
  struct fp96 {
     union {
         double f64;
         int64  i64;
     };
     union {
         float  f32;
         int32  i32;
     };
  };

For example:
  (gdb) info float
  ...
  f30: 0x41f8000041f00000 flt: 30                dbl: 6442451998.9999998
  f31: 0x403f000000000000 flt: 31                dbl: 31
  (gdb) p $f30
  $1 = {f32 = 30, f64 = 6442451998.9999998,
        i32 = 1106247680, i64 = 4753549407795806208}
  (gdb) p $f31
  $2 = {f32 = 31, f64 = 31,
        i32 = 1106771968, i64 = 4629418941960159232}

gdb/ChangeLog:

	* mips-tdep.c (CONF5_FRE): New define.
	(mips_set_float_regsize): Detect MIPS_FPU_HYBRID from Config5.FRE.
	(mips_float_regsize, mips_get_fp_single_location): Handle
	MIPS_FPU_HYBRID.
	(mips_get_fp_multi_location, mips_pseudo_register_read,
	mips_pseudo_register_write, mips_convert_register_p,
	mips_register_to_value, mips_value_to_register, mips_fp_type,
	mips_read_fp_register_single): Handle fp96.
	(mips_float_hybrid, mips_convert_register_float_fre_case_p,
	mips_fp96_type): New function.
	(mips_gdbarch_init): Initialise tdep->fp96_type.
	* mips-tdep.h (enum mips_fpu_mode): Add MIPS_FPU_HYBRID.
	(struct gdbarch_tdep): Update fp_mode comment, add fp96_type.