Release branch only commit: LDPC fixes
The assembler required LDPC to be 8 byte aligned when actually it
only needs to be 4-byte aligned with an 8-byte aligned label to
load from. The check is now relaxed to only check 4-byte alignment
as the linker may well misalign a label during linking anyway so
checking for 8-byte alignment is mostly pointless.