Disable checking for ISA mode jumps for the branch relocations.
We do however keep the micromips r6 branch relocations for the
bc and balc instructions.

This is due to a bug in the assembler that fails to mark labels
as being micromips if they appear at the end of the file or section.
Turing these checks back on needs to be thought about because
it could cause linker errors if existing objects/libraries (built
using the old assembler) are linked with the new linker.