Fix compatibility issue with PT_MIPS_ABIFLAGS handling
The PT_MIPS_ABIFLAGS header is unconditionally introduced if
a .MIPS.abiflags section is found.  This is OK for final linking
but it is possible to create an executable with a .MIPS.abiflags
section but without PT_MIPS_ABIFLAGS.  To create such a binary
in a real world environment then:

1) Take an object built with gas that has support for .MIPS.abiflags
2) Link this object using a linker that does not have support for
   .MIPS.abiflags
3) Objcopy or strip the resulting binary with a binutils that
   does have support for .MIPS.abiflags

The fix is to rely on objcopy's duplication of all PHDRs and not
introduce one when there is no link information. Existing entries
will therefore be retained but the PHDR section will not need
to grow.

bfd/
	* elfxx-mips.c (_bfd_mips_elf_additional_program_headers): Check
	if there is link information before adding a PT_MIPS_ABIFLAGS
	entry.
	(_bfd_mips_elf_modify_segment_map): Likewise.

ld/testsuite/

	* ld-mips-elf/abiflags-fake.d: New file.
	* ld-mips-elf/abiflags-fake.s: Likewise.
	* ld-mips-elf/abiflags-fake.ld: Likewise.
	* ld-mips-elf/mips-elf.exp: Run the new test.