Ensure that ASE information from the selected architecture is applied
DMZ2230 GAS does not enable implicit ASEs for most architectures
The rework of option handling done as part of .module implementation
left the implicit ASE logic broken and default enabled ASEs for
most architectures did not get applied.  This patch ensures the ASE
information is carried forward to the point where it is required.

Conflicts:
	gas/config/tc-mips.c