Disassembler fix.
Awkward fix for the bfd library. Previous versions of binutils would be
sloppy in copying metadata from one symbol to another. For microMIPS this
is terrible, a copied symbol would get the address with the LSB but would
not get the flag marking it as microMIPS. The disassembler and company when
iterating through a binary by symbols would then pick up the wrong addresses
and be unable to produce anything useful.

I'm calling this an awkward fix as we either lie outright to the user
or produce garbage results.