MIPS16E2 Assembler, Disassembler
An prototype assembler and disassembler for the experimental MIPS16E2 ASE. This
assembler is produced with the intention of measuring codesize,
opcodes/mips16-opc.c needs to be audited for correctness to ensure that operands
are not swapped unintentionally.

Currently produced object files are not marked as requiring MIPS16E2. Common ALU
operations are extended to 2 or 3 operand form. The input registers are the
traditional MIPS16 register set and the rest of the S registers. Output registers
are the traditional MIPS16 registers.

Opcode format for new extended operations:

EXTEND2 (5 bits) | Unused (5bits) | X (rx extension bit) | rN (4 bits) | Y (ry ext)

11111            | 00000          |                      |             |

Where X is the extension bit for rx in the RRR/RRI-A opcode space,
Y is the extension bit for ry for the RR opcode space.

The extension bit is the MSB of the register field.

rN is the additional register for any opcode space.