Release branch only commit: DMZ2367 ASEs are disabled with .set mips*
ASEs are supposed to be unaffected by a .set mips32r6 or similar but
all enabled ASEs are in fact disabled. This is both unintentional and
different behaviour compared to .set arch=mips32r6.

Ensure ASEs survive a change in architecture. This could result in
warnings about ASEs being disabled though.