uMIPS++ Add balc stub relaxation
Provisional implementation to analyze stub-relaxation on a per-object
basis. This will eventually be implemented by the linker and removed
from the assembler.

Optimization is enabled by default for R7. If required, disable with
-mno-balc-stubs option.

Known problem:
* size regressison: in some cases, where there are exactly 3 calls to
a function, gains made by relaxation are reversed when functions are
aligned to 4-bytes and we end up with a slightly bigger code section
than we started with.