Fix link-order resolution issues for IFUNC symbols
- Calculate iplt_offset from symbol index for shared global ifuncs
- Add dynamic tags to mark the start of .iplt (MIPS_IPLT) and first index
of the sequence of IFUNCs in dynsym table (MIPS_IFUNC_DYNINDX)
- Modify sort-logic of dynsym to sort all non-IFUNC symbols before IFUNCs
so that IFUNCs appear in one contiguous sequence of dynsym indices.
- Move iplt-stub creation to create_iplt from allocate_iplt so that
the correct stub offset is available for global shared symbols post-sort.
- Shuffled order of function definitions to avoid forward declarations

These changes solve the link-order resolution problem by allowing
unresolved ifunc symbol in an unrelocated load object to be resolved to the
corresponding .iplt stub.

Conflicts:

	bfd/elfxx-mips.c

Conflicts:

	bfd/elfxx-mips.c
	elfcpp/elfcpp.h