Comment clean-up and update XLP flag value
*** Skip ChangeLog

XLP flag value is updated so as to not collide with new extensions
introduced on other internal branches. References to microMIPS++
are changed to microMIPSr7 in all comments. The use of '++' is
still present in the opcode-table filename and there are multiple
occurences of micromipspp which may need refactoring later.

bfd/
	* reloc.c: Change in comments.

gas/
	* config/tc-mips.c: Change in comments.

opcodes/
	* micromips++-opc.c: Change in comments.

include/
	* elf/mips.h (AFL_ASE_XLP): Update flag value.
	(AFL_ASE_MASK): Likewise.
	* opcode/mips.h (ASE_XLP): Update flag value.