Bugzilla #2508 - segmentation fault in balc optimization
gas/
	* config/tc-mips.c (stubgroup_new): Forward declaration.
	(stub_funcless_mode): New.
	(procS): Move declaration forward.
	(cur_proc_ptr): Likewise.
	(md_assemble): Start new stub-group and toggle function-less mode.
	(balc_add_traverse): Add NULL-check.
	(s_change_sec): Terminate function-less stub group.
	(s_change_section): Likewise.
	(s_mips_ent): Terminate previous function-less stub group.
	(md_mips_ent): Likewise.

gas/testsuite/
	* gas/mips/balc-stub.s: New file.
	* gas/mips/balc-stub.l: New file.

*** Skip Changelog

This commit fixes problems with the previous attempt. It addresses
memory fault in #2525 relating to transition from a code section
to another code section. There is no functional change, except a
re-ordering of the stubgroup-handling in s_change_section() to
precede the early return statement. The test-case is updated to
exercise this hole.

gas/testsuite/
	* gas/mips/mips.exp: Add new test

Conflicts:

	gas/testsuite/gas/mips/balc-stub.s
	gas/testsuite/gas/mips/mips.exp