Bugzilla #2693: Disassemble ADDIUPC as LAPC by default
Change ADDIUPC to disassemble as LAPC by default. This is for
consistency between relocatable and linked objects. When an
addiupc has a alignment or size place-holder relocation, due to
link-relaxation, it will NOT disassemble correctly. The disassembler
expects a relocation on addiupc to be PC-relative, and if such
a relocation is not PC-relative then the interpretation of the
address field is incorrect.

opcodes/
	* nanomips-opc.c (nanomips_opcodes): Re-order to prefer
	LAPC over ADDIUPC.
	* nanomips-dis.c: (validate_insn_args): Don't match for
	relocation on ADDIUPC operand.

gas/testsuite/
	* gas/nanomips/nanomips.exp: Add test case.
	* gas/nanomips/nanomips-exc-noalias.d: New file.
	* gas/nanomips/nanomips-exc.d: Update reference output.
	* gas/nanomips/nanomips-exc32.d: Likewise.