Implement reconfigurable FPRs for Native debug
This patch implements the reconfigurable FPRs for Native debug scenario.
In GDB, 'target_ops' provides 'to_thread_architecture' which is supposed
to determine the architecture of the code where the target is currently
stopped at.

In this patch, MIPS native target overrides 'to_thread_architecture'
method which checks the FR bit of CP0 status register to decide if the
FPU mode has changed or not.  If it has changed then it requests the
target for the updated target description that matches the current FPU
mode, it then creates gdbarch accordingly and returns it.

'mips_linux_read_description' has been modified so that it always
returns a correct target description based on current FPU mode
(i.e according to state of FR bit of CP0 status register).

gdb/ChangeLog:

    * mips-linux-nat.c: Include "arch-utils.h", "sys/prctl.h".
    "sys/uio.h", "elf/common.h", "nat/linux-ptrace.h".
    "features/mips-fpu64-linux.c" and "features/mips-fpu64-dsp-linux.c".
    Delete includes: "nat/gdb_ptrace.h", <asm/ptrace.h>.
    (FIR_F64): New macro.
    (have_fp_mode): New variable.
    (mips_linux_get_fpu64): New function.
    (mips_linux_read_description): Probe for 64-bit vs 32-bit FPU.
    (mips_thread_architecture): New function.
    (_initialize_mips_linux_nat): Set mips_thread_architecture as
    to_thread_architecture routine.
    * mips-tdep.c (ST0_FR): Move macro to...
    * mips-tdep.h (STATUS_FR): ... this.

include/ChangeLog:
    * elf/common.h (NT_MIPS_FP_MODE): New macro.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Signed-off-by: Stefan Markovic <smarkovic@wavecomp.com>
Signed-off-by: Bhushan Attarde <bhushan.attarde@imgtec.com>