Implement reconfigurable FPRs for Native debug
This patch implements the reconfigurable FPRs for Native debug scenario.
In GDB, 'target_ops' provides 'thread_architecture' which is supposed
to determine the architecture of the code where the target is currently
stopped at.

In this patch, MIPS native target overrides 'thread_architecture' method
which checks if the FPU mode has changed or not. If it has changed then
it requests the target for the updated target description that matches
the current FPU mode, it then creates gdbarch accordingly and returns it.

gdb/ChangeLog:

	* mips-linux-nat.c (mips_linux_nat_target::thread_architecture):
	New function. If FPU mode changed reinitalize gdbarch via mips_gdbarch_init.
	* mips-tdep.c (mips_gdbarch_init): Use fpu_mode passed in via ...
	* mips-tdep.h (struct gdbarch_tdep_info): New struct definintion.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Signed-off-by: Stefan Markovic <smarkovic@wavecomp.com>
Signed-off-by: Bhushan Attarde <bhushan.attarde@imgtec.com>