Make gdbserver maintain process's target description dynamically
This patch makes GDBSERVER maintain the target description of the
process according to current state of process.

Current GDBserver tracks the XML target description per process.
It determines this description during start-up and then uses the same
description through out.  So, if the program changes it's state during
its execution such that it affects the target description (register
layout for example), then current GDBSERVER still uses the old target
description for the process.  That means, current GDBSERVER has no
provision to update the process's target description according to
process state.

This causes problems in case of MIPS considering a process can change
its FPU mode during the execution which causes the change in layout of
floating point registers (a switch from FR=0 to FR=1 causes FP register
size to be changed to 64-bit from 32-bit).

To handle this, this patch makes changes to GDBSERVER in that it always
look for updated target description while replying to
'qXfer:features:read' requests.  If it finds new description different
than that of the current one, then it makes new description as process's
target description and also rebuilds the regcache using this new
description.

For that purpose, 'fetch_target_desc' field is added to
'struct target_ops' and 'struct linux_target_ops' so that each target
and architecture can implement it accordingly.  This function is
supposed to return the correct target description matching the current
state of the program.  If not implemented then it defaults to return
"current_target_desc()".

For MIPS, existing "mips_read_description" function has been modified
to return correct target description matching current state of program
and has been set as 'fetch_target_desc' routine.

gdb/gdbserver/ChangeLog:

    * linux-low.c (linux_fetch_target_desc): New function.
    (linux_target_ops): Add linux_fetch_target_desc.
    * linux-low.h (linux_target_ops): New field fetch_target_desc.
    * linux-mips-low.c (tdesc_mips_fpu64_linux): New variable.
    (tdesc_mips_fpu64_dsp_linux): Likewise.
    init_registers_mips_fpu64_linux): New declaration.
    (init_registers_mips_fpu64_dsp_linux): Likewise.
    (STATUS_FR, FIR_F64): New defines.
    (MIPS64_EF_CP0_STATUS, MIPS64_ELF_NGREG): Likewise.
    (have_fpu64): New variable.
    (mips_read_description): Probe for 64-bit vs 32-bit FPU.
    (mips_fill_fpregset, mips_store_fpregset): Use $f0 rather than $0
    to determine FPR width.
    (the_low_target): Set mips_read_description as fetch_target_desc
    routine and set required fields to NULL.
    (initialize_low_arch): Add initializer functions for fpu64 and
    fpu64_dsp.
    * server.c (get_target_desc): New function.
    (get_features_xml): Use get_target_desc instead of
    current_target_desc.
    * target.h (target_ops): New field fetch_target_desc.
    * tdesc.c (compare_target_description): New function.
    * tdesc.h (compare_target_description): Declaration.