Workaround for line information in compressed MIPS functions
This fix allows the first address of a MIPS16/microMIPS function to
be used to get the filename/line and function that it belongs to
whereas it has previously needed the ISA bit to be set in the address
for the first instruction to match up correctly.

Note: we should really be checking the output of addr2line, but the
binutils test-framework doesn't have that capability yet.
Expected addr2line output for both new tests is:

$ addr2line -a 0x400000 0x400004 0x40000c -p -e <elf>
0x00400000: .*-lineno.c:1
0x00400004: .*-lineno.c:2
0x0040000c: .*-lineno.c:3

Changelog:

bfd/
	* elfxx-mips.c (_bfd_mips_elf_find_nearest_line): If no matching
	line is found for an address, retry with the ISA bit set in case
	searching at the beginning of a MIPS16/microMIPS section.

ld/testsuite/
	* ld-mips-elf/mips16-lineno.s: New test source.
	* ld-mips-elf/micromips-lineno.s: New test source.
	* ld-mips-elf/mips16-lineno.d: New test.
	* ld-mips-elf/micromips-lineno.d: New test.
	* ld-mips-elf/mips-elf.exp: Run the new tests

(cherry picked from commit 87678f2eba992f0306353ac81e7ed5bef7477b48)
(based on original commit by Matthew Fortune <matthew.fortune@mips.com>)