GTM18-262/PP118699: Fix addr2line mapping with garbage-collected modules
If all the functions in a C module are garbage-collected by the linker,
debug_ranges for that module consists of a list of empty ranges. In this
case, lookup falls to iterating over the line information without a range
match. If there is other valid code at lower addresses (starting at 0x0),
the defunct line information of a garbage-collected section will
accidentally match against that address, producing incorrect addr2line
output.

There is no clean way of fixing this, so we hack addr2line to distinguish
ranges that were truncated due to garbage-collection from ranges that
were not specified to begin with. If all range-information of a compilation
unit was truncated, then don't fall through to check line information, but
pass over that compilation unit.

bfd/
	* dwarf2.c (arange) <truncated>: New field.
	(arange_add): Set/reset truncated flag for the range.
	(comp_unit_find_nearest_line): Short-circuit to fail for
	compilation unit with only truncated ranges.