Subject: [PATCH 02/20] Add MIPS32 FPU64 GDB target descriptions
Add a couple of new target descriptions for MIPS32 targets with 64-bit
floating point registers, with and without DSP.

These are identical to their 32-bit FPU counterparts except that the FP
registers are 64-bits long to allow for debugging of F64=1 MIPS32
targets such as MIPS32r2 compatible cores running in FR=1 mode.

F64 targets have an FR bit in the status register to specify the
effective floating point register size.  When FR bit is 1 then the
register size is 64-bits and when FR bit is 0 then the register size
is 32-bit.

This patch also renames existing mips64-fpu.xml to mips-fpu64.xml so
that we can use a single template for the FPU, which is the same in
case of 64-bit FPU regardless of the machine word width of the CPU.

gdb/ChangeLog:

    * features/Makefile (WHICH): Add mips-fpu64-linux and
    mips-fpu64-dsp-linux.
    (mips-fpu64-expedite): Set.
    (mips-fpu64-dsp-expedite): Set.
    * features/mips64-fpu.xml: Rename to ...
    (features/mips-fpu64.xml): ... this.
    * features/mips64-dsp-linux.xml: Use renamed mips-fpu64.xml
    * features/mips64-linux.xml): Likewise.
    * features/mips-fpu64-dsp-linux.xml: New file.
    * features/mips-fpu64-linux.xml: New file
    * features/mips-fpu64-dsp-linux.c: New generated file.
    * features/mips-fpu64-linux.c: New generated file.
    * regformats/mips-fpu64-dsp-linux.dat: New generated file.
    * regformats/mips-fpu64-linux.dat: New generated file.

gdb/gdbserver/ChangeLog:

    * configure.srv (srv_regobj): Add mips-fpu64-linux.o and
    mips-fpu64-dsp-linux.o.
    (srv_xmlfiles): Add mips-fpu64-linux.xml,
    mips-fpu64-dsp-linux.xml and mips-fpu64.xml.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Signed-off-by: Bhushan Attarde <bhushan.attarde@imgtec.com>