Enhance cooked FP format
Define cooked FP types as unions representing the possible interpretations
of the registers.

Previously the type of the cooked floating point registers depended directly
on the raw type, so for MIPS32 targets they were single precision, and for
MIPS64 targets they were single or double precision depending on the
FP mode.

However in both cases it is possible for the FP registers to be interpreted
as either single or double precision values depending on the instruction
executed, and in a different way depending on the FP mode.

In case of double FPU --
When FR=0:
- Even FP registers can be interpreted as a double (aliasing both odd and
  even singles), or a single (aliasing the least significant half of the
  double).
- Odd FP registers can be interpreted only as a single (aliasing the most
  significant half of the even double).

When FR=1:
- All FP registers can be interpreted as a double, or a single (aliasing
  the least significant half of the double).

Some instructions can also treat them as 32-bit or 64-bit integers for the
purposes of converting between formats.

Therefore define three cooked fp types, as follows.

The first type, fp32_odd, is used for odd FP registers when FR=0, which
can only be interpreted as singles or 32-bit integers.  It is equivalent
to this C definition:
  union fp32_odd {
    float S;
    int32_t W;
  };

The second type, fp32_even, is used for even FP registers when FR=0,
which can be interpreted as either a single or double precision float,
or a 32-bit signed integer. The double-precision floating point (type D)
data type is stored in even-odd pairs of FPRs (so only even registers
can be interpreted as double precision floats), and the long-integer
(type L) and paired single (type PS) data types are not supported.
It is equivalent to this C definition:
  union fp32_even {
    float S;
    double D;
    int32_t W;
  };

The third type, fp64, is used for all FP registers when FR=1, which can
be interpreted as either singles or doubles, with the single always in
the least significant half of the double.  In this mode, Each FP
register is capable of storing any data type.  It is roughly
equivalent to this C definition (when interpreted as little endian):
  union fp64 {
    float  S;
    double D;
    float  PS[2];
    int32  W;
    int64  L;
    int32  PW[2];
  };

Note: `S',`D',`PS',`W',`L',`PW' are the architectural type designations
for various floating point formats.

Since multiple bits of code need to handle the different raw register
layouts and these new cooked types for different modes, the details are
abstracted with the following functions:
- mips_get_fp_single_location: Gets the raw register and offset containing
  a single precision register value.
- mips_get_fp_double_location: Gets the raw register(s) and offset(s)
  composing a double precision register value.
- mips_get_fp_multi_location: Gets the raw register(s) and offset(s)
  composing a cooked register value.

These abstractions are used by the code which reads and writes the cooked
(pseudo) FP registers, and the code which prints out the FP registers
("info float").

Now that the cooked type may be composed of multiple raw registers (where
only single precision raw values are provided), there is no need to
provide conversion functions from the 32-bit cooked value to double.
Instead the opposite is required for extracting a single out of the 64-bit
cooked value.

For example, with FR=0:
  (gdb) info float
  ...
  f0:  0x4131c28f flt: 11.1099997        dbl: 1163919.2546655273
  f1:  0x4131c28f flt: 11.1099997
  (gdb) p $f0
  $1 = {S = 11.1099997, D = 1163919.2546655273, W = 1093780111}
  (gdb) p $f1
  $2 = {S = 11.1099997, W = 1093780111}

And with FR=1:
  (gdb) info float
  ...
  f0:  0x40363851eb851eb8 flt: -3.21864398e+26   dbl: 22.219999999999999
  f1:  0x40363851eb851eb8 flt: -3.21864398e+26   dbl: 22.219999999999999
  (gdb) p $f0
  $3 = {S = -3.21864398e+26, D = 22.219999999999999, PS = {-3.21864398e+26,
    2.84718728}, W = -343597384, L = 4626947591664639672, PW = {-343597384,
    1077295185}}
  (gdb) p $f1
  $4 = {S = -3.21864398e+26, D = 22.219999999999999, PS = {-3.21864398e+26,
    2.84718728}, W = -343597384, L = 4626947591664639672, PW = {-343597384,
    1077295185}}

gdb/ChangeLog:

    * mips-tdep.c (mips_reg_part): New definition.
    (mips_get_fp_single_location, mips_get_fp_double_location,
    mips_get_fp_multi_location, mips_regcache_raw_read_parts
    mips_regcache_raw_write_parts, mips_fp32_odd_type,
    mips_fp32_even_type, mips_fp64_type, mips_fp_type): Likewise.
    (mips_pseudo_register_read, mips_pseudo_register_write,
    mips_register_type, mips_pseudo_register_type,
    mips_read_fp_register_single, mips_read_fp_register_double): Convert
    to handle new cooked FP format.
    (mips_gdbarch_init): Initialise tdep->fp32_odd_type,
    tdep->fp32_even_type and tdep->fp64_type.
    (mips_register_reggroup_p): Adjust.
    mips-tdep.h (struct gdbarch_tdep): Add fp32_odd_type, fp32_even_type
    and fp64_type.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Signed-off-by: Bhushan Attarde <bhushan.attarde@imgtec.com>