GTM19-502: Relax backend ABI check to conflicts from exact match
The nanoMIPS back-end verifies that the ABI in e_flags of a core-dump
matches the currently loaded executable, but linux doesn't set e_flags
at all. We relax the back-end check to only ensure that the ABI field
is not directly in conflict (i.e. p32 with p64 or vice-versa). This
allows the case of unset e_flags to just work.

bfd/
	* elfnn-nanomips.c (ABI_PNN_P): Remove macro.
	(ABI_P32_P): New macro.
	(ABI_P64_P): New macro.